third_party/blink/web_tests/wpt_internal/wake-lock/wakelock-onrelease.https.html

<!DOCTYPE html>
<link rel="help" href="https://w3c.github.io/wake-lock/#the-onrelease-attribute">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script>
// TODO(https://crbug.com/1015327): Move this to WPT once we can change
// permissions via testdriver.js.

async_test(async t => {
  window.testRunner.setPermission('wake-lock-screen', 'granted', location.origin, location.origin);

  const lock = await navigator.wakeLock.request("screen");
  lock.onrelease = t.step_func_done((ev) => {
    assert_class_string(ev, "Event", "release() must fire an Event object");
    assert_equals(ev.target, lock, "The event's target must be the lock that was acquired");
    assert_true(ev.isTrusted);
    assert_false(ev.bubbles);
    assert_false(ev.cancelable);
  });
  await lock.release();
}, "Test onreleased event's basic properties");
</script>
